+++ src/map_generator/StreetProperties.cpp
@@ -4,7 +4,7 @@
 
 const int StreetProperties::MAX_CAR_WIDTH = 2;
 const int StreetProperties::MAX_CARS_DISTANCE = 4;
-const int StreetProperties::LEVEL_VELOCITIES[] =
+const float StreetProperties::LEVEL_VELOCITIES[] =
 {
     1.f, 2.f, 3.f
 };
